LibWeb: Split PaintableWithLines from PaintableBox
