+++ Libraries/LibWeb/Painting/BackgroundPainting.cpp
@@ -15,7 +15,7 @@
 #include <LibWeb/Painting/Blending.h>
 #include <LibWeb/Painting/DisplayList.h>
 #include <LibWeb/Painting/DisplayListRecorder.h>
-#include <LibWeb/Painting/PaintableBox.h>
+#include <LibWeb/Painting/PaintableWithLines.h>
 
 namespace Web::Painting {
