+++ Libraries/LibWeb/Painting/PaintableBox.h
@@ -9,16 +9,12 @@
 
 #include <LibGfx/Forward.h>
 #include <LibWeb/CSS/StyleValues/GridTrackSizeListStyleValue.h>
-#include <LibWeb/Export.h>
-#include <LibWeb/Forward.h>
 #include <LibWeb/Layout/Box.h>
 #include <LibWeb/Painting/BackgroundPainting.h>
-#include <LibWeb/Painting/BorderPainting.h>
 #include <LibWeb/Painting/BoxModelMetrics.h>
 #include <LibWeb/Painting/ClipFrame.h>
 #include <LibWeb/Painting/Paintable.h>
 #include <LibWeb/Painting/PaintableFragment.h>
-#include <LibWeb/Painting/ShouldAntiAlias.h>
 
 namespace Web::Painting {
 
@@ -352,54 +348,5 @@ private:
 
     BoxModelMetrics m_box_model;
 };
-
-class PaintableWithLines : public PaintableBox {
-    GC_CELL(PaintableWithLines, PaintableBox);
-    GC_DECLARE_ALLOCATOR(PaintableWithLines);
-
-public:
-    static GC::Ref<PaintableWithLines> create(Layout::BlockContainer const&);
-    static GC::Ref<PaintableWithLines> create(Layout::InlineNode const&, size_t line_index);
-    virtual ~PaintableWithLines() override;
-
-    Vector<PaintableFragment> const& fragments() const { return m_fragments; }
-    Vector<PaintableFragment>& fragments() { return m_fragments; }
-
-    void add_fragment(Layout::LineBoxFragment const& fragment)
-    {
-        m_fragments.append(PaintableFragment { fragment });
-    }
-
-    virtual void paint(DisplayListRecordingContext&, PaintPhase) const override;
-
-    [[nodiscard]] virtual TraversalDecision hit_test(CSSPixelPoint position, HitTestType type, Function<TraversalDecision(HitTestResult)> const& callback) const override;
-
-    virtual void visit_edges(Cell::Visitor& visitor) override
-    {
-        Base::visit_edges(visitor);
-        for (auto& fragment : m_fragments)
-            visitor.visit(GC::Ref { fragment.layout_node() });
-    }
-
-    virtual void resolve_paint_properties() override;
-
-    size_t line_index() const { return m_line_index; }
-
-protected:
-    PaintableWithLines(Layout::BlockContainer const&);
-    PaintableWithLines(Layout::InlineNode const&, size_t line_index);
-
-private:
-    [[nodiscard]] virtual bool is_paintable_with_lines() const final { return true; }
-
-    Vector<PaintableFragment> m_fragments;
-
-    size_t m_line_index { 0 };
-};
-
-void paint_text_decoration(DisplayListRecordingContext&, TextPaintable const&, PaintableFragment const&);
-void paint_cursor_if_needed(DisplayListRecordingContext&, TextPaintable const&, PaintableFragment const&);
-void paint_text_fragment(DisplayListRecordingContext&, TextPaintable const&, PaintableFragment const&, PaintPhase);
-void paint_text_fragment_debug_highlight(DisplayListRecordingContext&, PaintableFragment const&);
 
 }
