+++ Libraries/LibWeb/Painting/SVGForeignObjectPaintable.h
@@ -8,7 +8,7 @@
 
 #include <LibGfx/ImmutableBitmap.h>
 #include <LibWeb/Layout/SVGForeignObjectBox.h>
-#include <LibWeb/Painting/PaintableBox.h>
+#include <LibWeb/Painting/PaintableWithLines.h>
 #include <LibWeb/Painting/SVGMaskable.h>
 
 namespace Web::Painting {
